+++ app/src/main/java/com/example/android/letseat/activities/mainMenu.java
@@ -1,23 +1,12 @@
-package com.example.android.letseat;
+package com.example.android.letseat.activities;
 
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.location.Location;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.widget.Toast;
 
-import com.google.android.gms.location.FusedLocationProviderClient;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-
-import java.util.Locale;
+import com.example.android.letseat.R;
 
 public class mainMenu extends AppCompatActivity {
