+++ app/src/main/java/com/example/android/letseat/utility/BusinessAdapter.java
@@ -1,4 +1,4 @@
-package com.example.android.letseat;
+package com.example.android.letseat.utility;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -9,6 +9,10 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
+
+import com.example.android.letseat.Business;
+import com.example.android.letseat.R;
+
 import java.util.ArrayList;
 
 public class BusinessAdapter extends ArrayAdapter<Business> {
