+++ app/src/main/java/com/example/android/letseat/utility/FetchDataAsyncTask.java
@@ -1,6 +1,10 @@
-package com.example.android.letseat;
+package com.example.android.letseat.utility;
 import android.os.AsyncTask;
 import android.util.Log;
+
+import com.example.android.letseat.App;
+import com.example.android.letseat.AsyncResponse;
+import com.example.android.letseat.R;
 
 import org.json.JSONArray;
 import org.json.JSONException;
