+++ app/src/main/res/layout/activity_business_list_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BusinessListView">
+    tools:context=".activities.BusinessListView">
 
     <ListView
         android:id="@+id/myListView"
