+++ app/src/main/res/layout/activity_surprise_me.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".surpriseMe">
+    tools:context=".activities.surpriseMe">
 
 
     <ProgressBar
@@ -16,7 +16,7 @@
         android:layout_height="wrap_content" />
 
     <fragment
-        class="com.example.android.letseat.BusinessDisplayFragment"
+        class="com.example.android.letseat.fragments.BusinessDisplayFragment"
         android:id="@+id/sm_fragment"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
