+++ app/src/main/res/layout/fragment_business_display.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".BusinessDisplayFragment">
+    tools:context=".fragments.BusinessDisplayFragment">
 
     <TextView
         style="@style/HeaderTextStyle"
